requirements.txt
azureml-core
python-dotenv
requests